refactoring: tests cleanup.
